refactor: move applyDepositForBuilder to processBuilderDepositRequest.ts
Co-locate the apply_deposit_for_builder helper with its only caller,
processBuilderDepositRequest, per review feedback. No behavior change: the
function and its util/gloas imports move verbatim; processDepositRequest drops
the now-unused gloas registry imports.
